refactor(xlsx): extract formula engine from renderer.ts
Move the ~600-line CF formula evaluator (tokenizer, recursive-descent
parser, A1/defined-name resolution, AND/OR/IF/ROUND/COUNTIF/SUMIF, and
date-serial helpers) into src/formula.ts. It is self-contained (depends
only on Cell/DefinedName types) and exposes evalFormulaToBool,
todaySerial, nowSerial.

renderer.ts drops from 4719 to 4118 lines. No behavior change: typecheck
passes and VRT regression is 65/65 at 0.0% diff.
